services: depend on network-target not interface-target
This change allow for deploying services that not just depend on the
network interfaces being configured but depend on the entire network
stack. Such as VPN interfaces, dns, ... .
My particular example is where I deploy these services listening on a
tinc vpn network interface, which is not configured until the
network-target stack completes, as it requires the network interfaces,
dns and so on to be available before the tinc vpn obtains an ip
address.